+++ src/handle-output.js
@@ -1,0 +1,39 @@
+const { DateTime } = require("luxon");
+
+function sendSwitch(node, onOff) {
+  const output1 = onOff ? { payload: true } : null;
+  const output2 = onOff ? null : { payload: false };
+  node.send([output1, output2, null]);
+}
+
+function runSchedule(node, schedule, time, currentSent = false) {
+  let remainingSchedule = schedule.filter((entry) => {
+    return DateTime.fromISO(entry.time) > time;
+  });
+  if (remainingSchedule.length > 0) {
+    const entry = remainingSchedule[0];
+    const nextTime = DateTime.fromISO(entry.time);
+    const wait = nextTime - time;
+    const onOff = entry.value ? "on" : "off";
+    node.log("Switching " + onOff + " in " + wait + " milliseconds");
+    const statusMessage = `${remainingSchedule.length} changes - ${
+      remainingSchedule[0].value ? "on" : "off"
+    } at ${nextTime.toLocaleString(DateTime.TIME_SIMPLE)}`;
+    node.status({ fill: "green", shape: "dot", text: statusMessage });
+    return setTimeout(() => {
+      sendSwitch(node, entry.value);
+      node.schedulingTimeout = runSchedule(node, remainingSchedule, nextTime);
+    }, wait);
+  } else {
+    const message = "No schedule";
+    node.warn(message);
+    node.status({ fill: "yellow", shape: "dot", text: message });
+    if (!currentSent) {
+      sendSwitch(node, node.outputIfNoSchedule);
+    }
+  }
+}
+
+module.exports = {
+  runSchedule,
+};
